# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: mini-evvy API
  description: |
    Event seating, bookings, payments, attendance, and background jobs.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /orgs/{orgId}/my-role:
    get:
      tags: [Organizations]
      summary: Get current user's role in organization
      operationId: getMyRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyRoleEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /orgs/{orgId}/members:
    post:
      tags: [Organizations]
      summary: Add member
      description: Requires org owner or admin. Role is `admin` or `member`.
      operationId: addMember
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    get:
      tags: [Organizations]
      summary: List members
      operationId: listMembers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/orgId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT from login/register. Subject is user UUID.
  parameters:
    orgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: "#/components/schemas/ErrorBody"
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: BOOKING_NOT_FOUND
        message:
          type: string
...
    AddMemberRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, member]
...
    Member:
      type: object
      properties:
        user_role_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        role_id:
          type: string
          format: uuid
        role_name:
          type: string
        organization_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
...
    MemberEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Member"
    MemberListEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/Member"
    MyRole:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [owner, admin, member]
    MyRoleEnvelope:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/MyRole"
...